Moved tests to other target
